+++ backend/src/main/java/com/quickbase/devint/service/resolver/CountryNameResolver.java
@@ -1,4 +1,4 @@
-package com.quickbase.devint.service.names;
+package com.quickbase.devint.service.resolver;
 
 public interface CountryNameResolver {
     String getOfficialName(String name);
