+++ backend/src/main/java/com/quickbase/devint/service/resolver/DefaultCountryNameResolver.java
@@ -1,4 +1,4 @@
-package com.quickbase.devint.service.names;
+package com.quickbase.devint.service.resolver;
 
 import java.util.Map;
